+++ wai-request-params/Test/Wai/Request/ParamsSpec.hs
@@ -59,9 +59,9 @@ spec = do
                 let (requestBody, request) = createRequestWithParams []
                 (paramOrNothing @UUID requestBody request "referredBy") `shouldBe` Nothing
 
-            it "should fail with a parser error on invalid input" $ do
+            it "should return Nothing on invalid input" $ do
                 let (requestBody, request) = createRequestWithParams [("referredBy", "not a uuid")]
-                (IO.evaluate (paramOrNothing @UUID requestBody request "referredBy")) `shouldThrow` (== ParamCouldNotBeParsedException { name = "referredBy", parserError = "has to be an UUID" })
+                (paramOrNothing @UUID requestBody request "referredBy") `shouldBe` Nothing
 
         describe "paramOrDefault" $ do
             it "should parse valid input" $ do
@@ -76,9 +76,9 @@ spec = do
                 let (requestBody, request) = createRequestWithParams []
                 (paramOrDefault @Int requestBody request 10 "page") `shouldBe` 10
 
-            it "should fail with a parser error on invalid input" $ do
+            it "should return default value on invalid input" $ do
                 let (requestBody, request) = createRequestWithParams [("page", "NaN")]
-                (IO.evaluate (paramOrDefault @Int requestBody request 10 "page")) `shouldThrow` (== ParamCouldNotBeParsedException { name = "page", parserError = "has to be an integer" })
+                (paramOrDefault @Int requestBody request 10 "page") `shouldBe` 10
 
 
         describe "paramList" $ do